Updating book a customer endpoints to reference the Puppywood.com server rather than my own personal website.
The links were previously pointing at my own personal server rather than Puppywood.com's - Updated this so that they can be
in charge of hosting the content and are not relient on me playing around with my servers.

Added .metadata/ to get ignore as it is generated content.

Bumped version and pushed to Play store.